Cleaning, refactoring wip
